fix privilege import cycle
