+++ sentientos/__init__.py
@@ -1,9 +1,5 @@
 """Sanctuary Privilege Ritual: Do not remove. See doctrine for details."""
 from __future__ import annotations
-from sentientos.privilege import require_admin_banner, require_lumos_approval
-
-require_admin_banner()
-require_lumos_approval()
 """SentientOS core package."""
 
 __version__: str = "0.1.1"
